requirements.txt
vk_api
pillow
vk